Fix guildRanks

+++ src/iologindata.cpp
@@ -504,7 +504,7 @@ bool IOLoginData::loadPlayer(Player* player, DBResult_ptr result)
 
 				if ((result = db->storeQuery(query.str()))) {
 					do {
-						guild->addRank(result->getNumber<uint32_t>("id"), result->getString("name"), result->getNumber<uint16_t>("level"));
+						guild->addRank(result->getNumber<uint32_t>("id"), result->getString("name"), result->getNumber<uint8_t>("level"));
 					} while (result->next());
 				}
 			}
@@ -518,7 +518,7 @@ bool IOLoginData::loadPlayer(Player* player, DBResult_ptr result)
 				query << "SELECT `id`, `name`, `level` FROM `guild_ranks` WHERE `id` = " << playerRankId;
 
 				if ((result = db->storeQuery(query.str()))) {
-					guild->addRank(result->getNumber<uint32_t>("id"), result->getString("name"), result->getNumber<uint16_t>("level"));
+					guild->addRank(result->getNumber<uint32_t>("id"), result->getString("name"), result->getNumber<uint8_t>("level"));
 				}
 
 				rank = guild->getRankById(playerRankId);
